requirements.txt
setuptools~=69.0.3
python-dotenv~=1.0.1
mpmath~=1.3.0
tabulate~=0.8.10